+++ src/PipeMvc/dotnetCampus.Ipc.PipeMvcServer/IpcFramework/IpcResponseMessageResult.cs
@@ -1,0 +1,14 @@
+﻿using dotnetCampus.Ipc.Messages;
+
+namespace dotnetCampus.Ipc.PipeMvcServer
+{
+    class IpcResponseMessageResult : IIpcResponseMessage
+    {
+        public IpcResponseMessageResult(IpcMessage responseMessage)
+        {
+            ResponseMessage = responseMessage;
+        }
+
+        public IpcMessage ResponseMessage { get; }
+    }
+}
